+++ Kernel/Output/HTML/Dashboard/TicketGeneric.pm
@@ -1159,7 +1159,13 @@ sub Run {
                 $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Pending till');
             }
             elsif ( $HeaderColumn eq 'CustomerCompanyName' ) {
-                $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer Company Name');
+                $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer Name');
+            }
+            elsif ( $HeaderColumn eq 'CustomerID' ) {
+                $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer ID');
+            }
+            elsif ( $HeaderColumn eq 'CustomerName' ) {
+                $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer User Name');
             }
             elsif ( $HeaderColumn eq 'CustomerUserID' ) {
                 $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer User ID');